+++ CHANGELOG.md
@@ -7,11 +7,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - Helpers
+    - `random-text`
+        - generates random text
     - `unique-id`
         - generate a unique-enough string for use in a DOM element's `id`
 - Tests
     - Integration
         - `unique-id`
+        - `random-text`
 - Ember Optional Features
     - `template-only-glimmer-components`
 - Mirage